HW-4: move instances
