+++ src/main/java/ru/gb/timesheet/page/TimesheetPageController.java
@@ -1,4 +1,4 @@
-package ru.gb.timesheet.controller;
+package ru.gb.timesheet.page;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -6,9 +6,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import ru.gb.timesheet.service.TimesheetPageService;
 
-import java.util.NoSuchElementException;
+import java.util.List;
 import java.util.Optional;
 
 
@@ -17,15 +16,21 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TimesheetPageController {
 
-    private final TimesheetPageService service;
+    private final TimesheetPageService timesheetService;
+
+    @GetMapping
+    public String getAllTimesheets(Model model) {
+        List<TimesheetPageDto> timesheets = timesheetService.findAll();
+        model.addAttribute("timesheets", timesheets);
+        return "timesheets-page.html";
+    }
 
     @GetMapping("/{id}")
     public String getTimesheetPage(@PathVariable Long id, Model model) {
-        Optional<TimesheetPageDto> timesheetPageDto = service.findById(id);
+        Optional<TimesheetPageDto> timesheetPageDto = timesheetService.findById(id);
 
         if (timesheetPageDto.isEmpty()) {
-            //return not-found.html
-            throw new NoSuchElementException();
+            return "not-found.html";
         }
 
         model.addAttribute("timesheet", timesheetPageDto.get());
