+++ src/main/java/ru/gb/timesheet/page/TimesheetPageDto.java
@@ -1,4 +1,4 @@
-package ru.gb.timesheet.controller;
+package ru.gb.timesheet.page;
 
 import lombok.Data;
 
@@ -6,6 +6,7 @@ import lombok.Data;
 public class TimesheetPageDto {
 
     private String projectName;
+    private String projectId;
     private String id;
     private String minutes;
     private String createdAt;
